Minor code adjustments

+++ osu.Game/Beatmaps/BeatmapStatistic.cs
@@ -18,7 +18,8 @@ namespace osu.Game.Beatmaps
 
         public string Content;
         public LocalisableString Name;
-        public float Count;
-        public float Maximum;
+
+        public int Count;
+        public int Maximum;
     }
 }
